Fix enum flags

+++ Source/Core/Meta/enum.h
@@ -37,8 +37,8 @@ inline constexpr T CombineFlags(T&& flags...) {
 //----------------------------------------------------------------------------
 #define ENUM_FLAGS(_ENUMTYPE) \
     STATIC_ASSERT(std::is_enum<_ENUMTYPE>::value); \
-    inline _ENUMTYPE operator &(_ENUMTYPE lhs, _ENUMTYPE rhs) { return _ENUMTYPE(lhs&rhs); } \
-    inline _ENUMTYPE operator |(_ENUMTYPE lhs, _ENUMTYPE rhs) { return _ENUMTYPE(lhs|rhs); }
+    inline _ENUMTYPE operator &(_ENUMTYPE lhs, _ENUMTYPE rhs) { return _ENUMTYPE(u64(lhs)&u64(rhs)); } \
+    inline _ENUMTYPE operator |(_ENUMTYPE lhs, _ENUMTYPE rhs) { return _ENUMTYPE(u64(lhs)|u64(rhs)); }
 //----------------------------------------------------------------------------
 //////////////////////////////////////////////////////////////////////////////
 //----------------------------------------------------------------------------
